backend/requirements.txt
Django>=6.0,<7.0
djangorestframework>=3.16,<4.0
django-cors-headers>=4.7,<5.0
python-decouple==3.8
dj-database-url>=2.3,<3.0
psycopg2-binary>=2.9,<3.0
gunicorn>=23.0,<24.0
